update annotation works (PUT method)
git-svn-id: https://svn.clarin.eu/DASISH/t5.6/backend/annotator-backend/trunk/annotator-backend@3906 72633d89-856b-0410-a642-59178932f03a

+++ src/main/java/eu/dasish/annotation/backend/dao/impl/DBIntegrityServiceImlp.java
@@ -368,7 +368,10 @@ public class DBIntegrityServiceImlp implements DBIntegrityService {
     // TODO: unit test
     @Override
     public int updateUsersAnnotation(Number userID, Annotation annotation) throws SQLException, Exception {
-        int updatedAnnotations = annotationDao.updateAnnotation(annotation, userID);
+        if (userID != userDao.getInternalIDFromURI(annotation.getOwnerRef())) {
+            throw new Exception("The current user is not allowed to update this annptation, because (s)he is not its owner");
+        }
+        int updatedAnnotations = annotationDao.updateAnnotation(annotation, userID);        
         Number annotationID = annotationDao.getInternalIDFromURI(annotation.getURI());
         int deletedTargets = annotationDao.deleteAllAnnotationTarget(annotationID);
         int deletedPrinsipalsPermissions = annotationDao.deleteAnnotationPrincipalPermissions(annotationID);
